requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
requests==2.31.0
google-generativeai==0.4.1
python-dotenv==1.0.1
flask-mail==0.9.1